RandomQuantum Documentation

Functions

More info to come

Data Structs

RandomQuantum.GinibreEnsemble — Type

RandomQuantum.GinibreEnsemble(rows,cols)

Type corresponding to the Ginibre distribution of complex matrices.

RandomQuantum.FubiniStudyPureState — Type

RandomQuantum.FubiniStudyPureState(dim)

Type corresponding to unitarily-invariant distribution of pure states for some Hilbert space dimension.

RandomQuantum.FubiniStudyMixedState — Type

RandomQuantum.FubiniStudyMixedState(dim, bath_dim)

Type corresponding to distribution of mixed states obtained by tracing out part of a pure state obtained from a FubiniStudy distribution. When dim == bath_dim, this is identical to the Hilbert-Schmidt distribution of mixed states. See, e.g., Zyczkowski et al., J. Math. Phys. 52, 062201 (2011) arXiv:1010.3570.

RandomQuantum.HilbertSchmidtMixedState — Type

RandomQuantum.HilbertSchmidtState(dim)

Type corresponding to distribution of mixed states according to the Hilbert-Schmidt measure (induced by the Frobenius distance). This is identical to the mixed state distribution induced by tracing out half of a pure state distributed according to the Fubini-Study metric. See, e.g., Zyczkowski et al., J. Math. Phys. 52, 062201 (2011) arXiv:1010.3570.

RandomQuantum.BuresMixedState — Type

RandomQuantum.BuresMixedState(dim)

Type corresponding to distribution of mixed states according to the Bures metric. See, e.g., Zyczkowski et al., J. Math. Phys. 52, 062201 (2011) arXiv:1010.3570.

RandomQuantum.ClosedHaarEnsemble — Type

RandomQuantum.ClosedHaarEnsemble(dim)

Type corresponding to the unitarily invariant distribution of unitary transformations for some Hilbert space dimension. See, e.g., Mezzadri, Notices Amer Math Soc 54 4 592 (2007).

RandomQuantum.OpenHaarEnsemble — Type

RandomQuantum.OpenHaarEnsemble(dim, bath_dim)

Type corresponding to the unitarily invariant distribution of unitary transformations for system and bath, such that the bath (initially in the ground state) is traced out. See, e.g., Bruzda et al., Phys. Lett. A 373, 320-324 (2009), arXiv:0804.2361.

Missing docstring.

Missing docstring for GUE. Check Documenter's build log for details.

RandomQuantum.RandomClosedEvolution — Type

RandomQuantum.RandomClosedEvolution(dim, α)

Type corresponding to the integrated evolution of random Hamiltonians (with unitarily invariant distribution) on some Hilbert space.

RandomQuantum.RandomOpenEvolution — Type

RandomQuantum.RandomOpenEvolution(dim, bath_dim, α)

Type corresponding to the distribution of completely positive trace preserving maps obtained from the integrated evolution of random Hamiltonians (with unitarily invariant distribution) on a dimensional Hilbert space, such that bath (initially in the ground state) is traced out after the evolution.

Index

RandomQuantum.BuresMixedState

RandomQuantum.ClosedHaarEnsemble

RandomQuantum.FubiniStudyMixedState

RandomQuantum.FubiniStudyPureState

RandomQuantum.GinibreEnsemble

RandomQuantum.HilbertSchmidtMixedState

RandomQuantum.OpenHaarEnsemble

RandomQuantum.RandomClosedEvolution

RandomQuantum.RandomOpenEvolution